Added the rest of the content
I added the cards, reviews, help cards, announcements and the footer.
I styled them and made them look as good on different screen sizes with mobile first approach in mind - still have some styling to do but looks not too bad currently
Had some issues with the borders for the help cards and the review but managed to get them look as intended.
Added the footer and divided it into 2 sections. Then I styled them separately as it is in the concept image
Used display-flex for all my elements in order to make them behave as intended. I changed the flex-direction on a few places for small screens