Feedback
I fixed all the issues I had for all screen sizes
I asked for feedback and a developer was kind enough to provide some. I have fixed the issues raised:
I made login and primary buttons links and cursor appears when hovering over them
The testimonial was wider than the elements above and below. I restricted its width in conjunction with the cards so they have same width on small screens
Social logos were set as flex-direction: column; for small screens but there is space and does not look good. I changed the flex-wrap property to wrap in order to allow them to go in column when the screen shrinks. That way they are side by side until the screen allows.
The footer links did not look nice on small screens. I changed the flex-direction to column for small screens and added some padding to make them look good.
Added some padding to the nav links on small screens as they were stuck to each other